remove input.js, move functionalities to their own domains
